+++ sample.env
@@ -3,3 +3,5 @@ SECRET_TOKEN=misskey_token
 
 CONFIG_DIR=./config
 RUN_SERVER=false
+DB_TYPE=pickle
+DB_URL=ssss
